app/requirements.txt
Flask==0.10.1
gunicorn==19.4.5
PyGithub
celery==3.1.20
redis==2.10.5